+++ quil-py/src/instruction/measurement.rs
@@ -6,13 +6,13 @@ use rigetti_pyo3::{
 };
 
 use crate::{
-    impl_copy_for_instruction, impl_eq, impl_to_quil,
+    impl_copy_for_instruction, impl_eq, impl_pickle_for_instruction, impl_to_quil,
     instruction::{PyMemoryReference, PyQubit},
 };
 
 py_wrap_data_struct! {
-    #[pyo3(subclass)]
     #[derive(Debug, PartialEq, Eq)]
+    #[pyo3(subclass, module = "quil.instructions")]
     PyMeasurement(Measurement) as "Measurement" {
         qubit: Qubit => PyQubit,
         target: Option<MemoryReference> => Option<PyMemoryReference>
@@ -23,6 +23,7 @@ impl_copy_for_instruction!(PyMeasurement);
 impl_hash!(PyMeasurement);
 impl_repr!(PyMeasurement);
 impl_eq!(PyMeasurement);
+impl_pickle_for_instruction!(PyMeasurement);
 
 #[pymethods]
 impl PyMeasurement {