+++ quil-py/src/instruction/mod.rs
@@ -1,3 +1,6 @@
+use std::str::FromStr;
+
+use pyo3::exceptions::PyValueError;
 use quil_rs::instruction::Instruction;
 use rigetti_pyo3::{
     create_init_submodule, impl_repr, py_wrap_union_enum,
@@ -15,8 +18,8 @@ pub use self::{
     circuit::PyCircuitDefinition,
     classical::{
         PyArithmetic, PyArithmeticOperand, PyArithmeticOperator, PyBinaryLogic, PyBinaryOperand,
-        PyBinaryOperands, PyBinaryOperator, PyComparison, PyComparisonOperand,
-        PyComparisonOperator, PyConvert, PyExchange, PyMove, PyUnaryLogic, PyUnaryOperator,
+        PyBinaryOperator, PyComparison, PyComparisonOperand, PyComparisonOperator, PyConvert,
+        PyExchange, PyMove, PyUnaryLogic, PyUnaryOperator,
     },
     control_flow::{PyJump, PyJumpUnless, PyJumpWhen, PyLabel, PyTarget, PyTargetPlaceholder},
     declaration::{
@@ -108,6 +111,14 @@ impl PyInstruction {
         self.as_inner().is_quil_t()
     }
 
+    #[staticmethod]
+    pub fn parse(string: &str) -> PyResult<Self> {
+        match Instruction::from_str(string) {
+            Ok(instruction) => Ok(Self(instruction)),
+            Err(err) => Err(PyValueError::new_err(err.to_string())),
+        }
+    }
+
     // Implement the __copy__ and __deepcopy__ dunder methods, which are used by Python's
     // `copy` module.
     //
@@ -140,7 +151,6 @@ create_init_submodule! {
         PyArithmeticOperator,
         PyBinaryLogic,
         PyBinaryOperand,
-        PyBinaryOperands,
         PyBinaryOperator,
         PyComparison,
         PyComparisonOperand,