+++ quil-rs/src/parser/lexer/mod.rs
@@ -17,7 +17,7 @@ mod quoted_strings;
 mod wrapped_parsers;
 
 use nom::{
-    bytes::complete::{is_a, is_not, take_while, take_while1},
+    bytes::complete::{is_a, take_till, take_while, take_while1},
     character::complete::{digit1, one_of},
     combinator::{all_consuming, map, recognize, value},
     multi::many0,
@@ -183,7 +183,7 @@ fn lex_data_type(input: LexInput) -> InternalLexResult {
 
 fn lex_comment(input: LexInput) -> InternalLexResult {
     let (input, _) = tag("#")(input)?;
-    let (input, content) = is_not("\n")(input)?;
+    let (input, content) = take_till(|c| c == '\n')(input)?;
     Ok((input, Token::Comment(content.to_string())))
 }
 
@@ -297,7 +297,11 @@ fn lex_number(input: LexInput) -> InternalLexResult {
     Ok((
         input,
         match integer_parse_result {
-            Ok(_) => Token::Integer(float_string.parse::<u64>().unwrap()),
+            Ok(_) => float_string
+                .parse::<u64>()
+                .map(Token::Integer)
+                .map_err(|e| InternalLexError::from_kind(input, e.into()))
+                .map_err(nom::Err::Failure)?,
             Err(_) => Token::Float(double(float_string)?.1),
         },
     ))
@@ -389,14 +393,18 @@ mod tests {
 
     #[test]
     fn comment() {
-        let input = LocatedSpan::new("# hello\n#world");
+        let input = LocatedSpan::new("# hello\n#world\n#\n#");
         let tokens = lex(input).unwrap();
         assert_eq!(
             tokens,
             vec![
                 Token::Comment(" hello".to_owned()),
                 Token::NewLine,
-                Token::Comment("world".to_owned())
+                Token::Comment("world".to_owned()),
+                Token::NewLine,
+                Token::Comment("".to_owned()),
+                Token::NewLine,
+                Token::Comment("".to_owned())
             ]
         )
     }